+++ Mixture/src/Platform/Vulkan/Buffer/IndexBuffer.hpp
@@ -10,7 +10,7 @@ namespace Mixture::Vulkan
     class IndexBuffer : public Mixture::IndexBuffer
     {
     public:
-        IndexBuffer(const std::vector<uint16_t>& indices);
+        IndexBuffer(const std::vector<uint32_t>& indices);
         ~IndexBuffer() override;
         
         void Bind(CommandBuffer commandBuffer) override;
